+++ common/src/main/java/com/kryeit/missions/mission_types/create/train/TrainDriverPassengerMission.java
@@ -1,5 +1,6 @@
 package com.kryeit.missions.mission_types.create.train;
 
+import com.kryeit.JSONObject;
 import com.kryeit.Missions;
 import com.kryeit.missions.MissionDifficulty;
 import com.kryeit.missions.MissionManager;
@@ -11,18 +12,41 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.UUID;
 
-public class TrainDriverMissionType implements MissionType {
+public class TrainDriverPassengerMission implements MissionType {
     private static final ResourceLocation IDENTIFIER = new ResourceLocation(Missions.MOD_ID, "distance");
+    private static final ResourceLocation PASSENGERS = new ResourceLocation(Missions.MOD_ID, "passengers");
 
     public static void handleDistanceChange(UUID player, int difference) {
-        MissionManager.incrementMission(player, TrainDriverMissionType.class, IDENTIFIER, difference);
+        MissionManager.incrementMission(player, TrainDriverPassengerMission.class, IDENTIFIER, difference);
+    }
+
+    public static int passengersNeeded() {
+        int passengers = 1;
+        try {
+            File file = Path.of("missions").toFile();
+            if (!file.exists()) return passengers;
+            String jsonContent = new String(Files.readAllBytes(Paths.get("missions/missions.json")));
+            JSONObject object = new JSONObject(jsonContent);
+            JSONObject thing = object.getObject("train-driver-passenger");
+            if (thing != null) {
+                passengers = Integer.parseInt(thing.getObject("requirements").getString(PASSENGERS.toString()));
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return Math.max(passengers, 1);
     }
 
     @Override
     public String id() {
-        return "train-driver";
+        return "train-driver-passenger";
     }
 
     @Override
@@ -32,12 +56,12 @@ public class TrainDriverMissionType implements MissionType {
 
     @Override
     public MissionDifficulty difficulty() {
-        return MissionDifficulty.NORMAL;
+        return MissionDifficulty.HARD;
     }
 
     @Override
     public Component description() {
-        return Component.nullToEmpty("Train driving mission");
+        return Component.nullToEmpty("Train passenger mission");
     }
 
     @Override
@@ -62,7 +86,7 @@ public class TrainDriverMissionType implements MissionType {
 
     @Override
     public ItemStack getPreviewStack(ResourceLocation item) {
-        return AllBlocks.TRAIN_CONTROLS.asStack();
+        return AllBlocks.TRACK_STATION.asStack();
     }
 
 }